Update: Implemented new firmware update command $27
This command accepts an arbitrary url pointing an esp32 firmware image.
Since this is different from the way the legacy update commands worked,
this command uses the previously unused id $27. The legacy update commands
will need to be deprecated with an appropriate log message.

See comments in update.cpp for details on how this command works.