Update: Reimplemented and improved update.php
The update.php script will now only report current and previous
unstable versions if the unstable versions found in the directory are
actually based on the current stable version. It would not make sense
to downgrade to an unstable 2.0.0-23 once a new stable 2.1.0 has
been officially released, because the new stable will include all changes
included in the previous unstable releases.