Update.asm: Added warning/confirmation for unstable versions, improvements
Before installing unstable versions a warning is issued and the user is
asked to confirm the installation.

Properly format error messages received from the update command (ascii).

Scan for exit keys run/stop and back-arrow from every prompt so that the
user can always exit to portal (actual return not implemented yet).

Changed online-flasher url to the the designated location.