Deprecated: Implement command to log/reply properly to deprecated commands
This commands handles all command ids deprecated with version 2.0.0.

An appropriate message containing the reason for the deprecation as well
as reference to new commands implementing similar functionality is
generated and logged to the console as well as send as the command reply.

This should make it more obvious to developers stumbling over these old
commands. If they are not watching the logs, they might at least discover
the message in the command response.