Undefined: Implemented undefined command to handle undefined command ids
Prior to this, the Service component would simply abort early when
detecting an undefined command id, letting the client run into a timeout.

Now the Undefined command handles undefined ids by logging to the console
but also sending a proper error message back to the client.

This way developers that don't watch the serial console get a hint on the
client as to why the command had no effect, e.g. the response may read
"Undefined command id 0x42".