Update: Improved update procedure, added explicit Reboot command
The update command will no longer reboot immediately after flashing into
the OTA partition, so the command always send a proper reply to the client.

Upon successful completion, the client can instead use the new Reboot
command to boot into the newly installed firmware. This command does not
send a reply before rebooting, but instead sets a flag in flash. This
flag is checked in Userport::connect() on boot. If set, a single handshake
is send to the client to confirm successful reboot, and the flag is cleared
again.

This way the update program can properly wait for the reboot to complete.

Added appropriate public methods to Userport to send handshakes before
and after reboot. When the reboot aborts the command by rebooting the ESP,
the clients sending routine is still waiting for the last byte of the
request to be acknowledged with a handshake, thus reboot sends this
handshake immediately before restarting the ESP. The client then expects
the confirming handshake after reboot using wic64_wait_for_handshake.