Service/Command: Added support for revised request header format
Added support for a revised request header format using the new api id
$52 (ascii "R"). The new request header format is

"R", <command>, <payload-size-low>, <payload-size-high>, [<payload>...]

In contrast to the legacy command header format, payload size is now
specified after the command id byte and contains the actual size of the
payload itself, instead of the size of the entire request.

In addition, for commands using the revised header format, the response
size for http requests is no longer adjusted by substracting two bytes
if the request url ends with ".prg". This was a hack included in the
original firmware for unknown reasons.

For commands using the revised header format, the response size is now
transferred in little-endian byte order instead of big-endian.

The legacy request header format is still supported and will continue to
be supported in the foreseeable future.