HttpClient: Removed explicit handling of keep-alive connections
I was previously unaware of the fact that for HTTP/1.1, keep-alive is
always assumed, and the Keep-Alive header is only required to negotiate
keep-alive for HTTP/1.0.

Removed explicit checks for the Keep-Alive header and simply assume that
connections can be reused. If a connection can not be reused, the existing
code already opens a new connection anyways.