Data/Commands: Add terminating nullbyte to all proper string responses
In order to make life easier on the client side, all responses that
constitute informative strings that are likely to be output to the user
are transferred with an additional terminating null byte. This ways the
client can just pass the reponse pointer to a print routine like $ab1e
without having to null the response buffer in advance or having to attach
a nullbyte manually.

These include the ip, mac, ssid and rssi as well as informative messages.
Only strings that are used as success or error codes, such as "!0", "!E"
or "0" are still send without a terminating null byte since it may be
sufficient to simply test the response size to detect an error response,
as it is already done in existing client code.

The Data method wrap() has been renamed to set() while the former
copy(char*) method has been renamed to copyData(char*). In addition, the
new method copyString(char*) now copies the string contents and adds
an additional nullbyte to the response data.