HttpClient: Always close connection when reading queued response fails
If queueing the response of an extended HTTP request fails (read error
or queue timeout), always close the current connection. This is crucial
because otherwise the request will stay open until the server times out,
and an immediate subsequent attempt to reuse the connection take too much
time until the first retry, causing the client to run into a timeout.