Connected: Added new command Connected (0x2c) with esp side timeout
This commands waits for a WiFi connection to become avaible for the number
of seconds specified in the first byte of the payload. If a connection is
already established or becomes available during this time, the command
will return immediately with a success status code. If no usable connection
is available after the specified number of seconds, a CONNECTION_ERROR is
reported and the status message will contain further details (either there
is no connection available, or the connection is established but no IP
address has been assigned).