Timeout: Added command Timeout ($2d) to set ESP timeout for next request
This command sets the ESP timeout to use for data transfers and queued
sends/receives to/from remote servers. It takes a single payload byte
specifiying the timeout to use in seconds. The timeout set by this
command will only be used during the following request and will be reset
to the  default value of 1000ms afterwards, regardless of whether the
request completes successfully or is aborted due to an error condition.
This measure ensures that the ESP timeout is always in a defined state,
e.g. not left at an unexpected value by a previously running program.

This command allows programs that send or receive in discrete steps to
control how long the ESP will wait for the transfer to continue. For
example, a program that receives data in chunks and needs more than one
second to processes each chunk (e.g. for writing data to disk) can adjust
the ESP timeout accordingly and prevent the ESP from prematurely aborting
the transfer.