Http: Corrected handling of the legacy 201 status code mechanism
Renamed endPositiveResponseOnStatusCode201() to
sendRequestedResponseOnStatusCode201() and send the value of the third
field of the server response as the command response, as it was
implemented in the legacy firmware, and no longer send "00" by default.

This is only required since there is still client code that expects the
respective response when calling certain URLs on the portal server.