Userport: Fixed handling of last handshake, improve request frequency
When requests were made by the client in quick succession, the intial
handshake was missed sporadically, especially on higher log levels.

This was due to subtle timing issues when trying to handle the last
handshake of the previous request in onTransferAborted(), where the final
handshake is send to the client. Trying to handle this in the
ISR as well as the onTransferAborted() method running in the Userport
event task proved to be tricky because the method had to assume that the
ISR handled the TRANSFER_STATE_TERMINATING after an artificial delay.

Now the last handshake is simply handled in onRequestInitiated(). If an
initial handshake is received while the data direction line is still HIGH,
we can safely assume that this is handshake was caused by reading the
last byte of the previous request, and not an attempt to start a new
request.

This also made it possible to reduce the artificial delays before/after
sending transitional handshakes from 65ms to 7ms per transfer, increasing
the maximum frequency at which requests can be send. This allows for
faster active polling, e.g. when polling a TCP connection for data.

In addition the data direction line is no longer handled by an ISR, which
means that aborting a tranfer session before completing the previous
session is no longer supported. This means that in theory the PA2 line can
be ignored, but for now the firmware still checks if the line is in the
expected state when before receiving/sending.

Removed displaying the amount of memory after every transfer since it
seems the display update is too expensive if requests are send at high
frequencies.