Service: Don't accept new request if a request is still being serviced
This may happen if client code immediately retries to send a request if
the previous request times out while Service is "in between" transfers,
e.g. in the process of making a remote request. In this case an
informative error message is logged, suggesting to increase the client
side timeout.

Problems like this could be avoided if the client-side timeout was sent
with the request, and the server side timeout is kept slightly below the
client side timeout. This would make sure that client code is always safe
to retry right away since the server side would always timeout ahead of
the client.