Userport: Pull handshake line low for at least 5us
Without this delay, it seems 469 boards (or specific CIA revisions?) don't
see the handshake from the ESP.

This delay was present in the legacy firmware as well, but I removed it
since it didn't seem to matter. Turns out the delay is not required on
the 466 board I used, but might be needed for other boards or for specific
CIA revisions.

I'll do some more testing to determine whether this is caused by the CIA
revision used or by differences between board revisions.