Commands: Added WIC64_CMD_FORCE_TIMEOUT ($fc), WIC64_CMD_FORCE_ERROR ($fd)
Added commands to test timeout and error handling.

WIC64_CMD_FORCE_TIMEOUT will delay for the number of seconds specified in
the first byte of the payload. If no payload is sent, the delay defaults
to one second.

WIC64_CMD_FORCE_ERROR will always return an INTERNAL_ERROR with the status
message "Test error".

Renamed Echo command class to Test and handle the new commands in it as
well.