Url: Reverted changes to substitution of %mac back to legacy behavior
Reverted changes to the substitution of %mac placeholder in URLs back to
the behavior of the legacy firmware, i.e. substitute %mac with MAC-Address
string followed by the user's session id string.

This prevents breaking third party application that use the substituted
string to identify users who are logged into the portal.